requirements_test.txt
opencv-contrib-python
ffmpeg-python
pytest
pytest-sugar
